requirements/base.txt
git+https://github.com/luizalabs/django-postgreconnect.git
dj-database-url==0.5.0
Django==3.1.2
django-cid==2.1
django-filter==2.4.0
django-session-timeout==0.1.0
djangorestframework==3.12.1
djangorestframework-filters==1.0.0.dev2
djangorestframework-xml==2.0.0
drf-yasg==1.20.0
gunicorn==20.0.4
httptools==0.1.1
ipython==7.19.0
markdown==3.3.3
psycopg2-binary==2.8.6
simple-settings==1.0.0
structlog==20.1.0
uvicorn==0.12.2
uvloop==0.14.0
whitenoise==5.2.0
